fix(server): handle all IfcProperty subtypes for full parity; split data_model tests (ratchet)
Adversarial review (MERGE-READY) flagged that only IFCPROPERTYSINGLEVALUE was
resolved server-side, so enumerated/bounded/list/table/reference properties —
common in real files (e.g. an enumerated AcousticRating) — stayed blank on the
server path while the WASM path shows them, contradicting "identical". Extend
`extract_property` to mirror every `parsePropertyValue` branch's display value:
- ENUMERATEDVALUE / LISTVALUE → members joined ", "
- BOUNDEDVALUE → "value [lower – upper]" (kind kept `string` so the client
  doesn't Number() the display to NaN; measure tag on data_type)
- TABLEVALUE → "Table (N rows)"; REFERENCEVALUE → "#id"
The cross-path parity test + Rust test now cover an enumerated value (`R1, R2`
identical on both paths).

Also splits the data_model + parquet_data_model inline test modules into
exempt `tests.rs` / `_tests.rs` files so the module-size ratchet passes without
raising any budget (mod.rs 480→241, parquet 763→660).
