+++ apps/server/src/services/data_model/properties.rs
@@ -71,13 +71,79 @@ pub(super) fn extract_properties(
 /// code only matched bare `String`/`Float` and emitted `format!("{:?}")` Debug
 /// garbage for every text/boolean value — this resolves them properly.
 fn extract_property(entity: &DecodedEntity, _decoder: &mut EntityDecoder) -> Option<Property> {
-    // IfcPropertySingleValue: [0]=Name, [1]=Description, [2]=NominalValue, [3]=Unit
-    if !entity.ifc_type.as_str().eq_ignore_ascii_case("IFCPROPERTYSINGLEVALUE") {
-        return None;
-    }
+    use ifc_lite_core::AttributeValue;
+    // All IfcProperty subtypes carry Name at attribute 0.
     let property_name = entity.get_string(0)?.to_string();
-    let nominal_value = entity.get(2)?;
-    let (property_value, property_type, data_type) = resolve_single_value(nominal_value);
+    let ty = entity.ifc_type.as_str().to_uppercase();
+
+    let (property_value, property_type, data_type) = match ty.as_str() {
+        // [Name, Description, NominalValue, Unit]
+        "IFCPROPERTYSINGLEVALUE" => resolve_single_value(entity.get(2)?),
+
+        // [Name, Description, EnumerationValues (list), EnumerationReference]
+        "IFCPROPERTYENUMERATEDVALUE" => {
+            let joined = join_member_list(entity.get(2));
+            (joined.unwrap_or_default(), "string".into(), None)
+        }
+
+        // [Name, Description, ListValues (list), Unit]
+        "IFCPROPERTYLISTVALUE" => {
+            let joined = join_member_list(entity.get(2));
+            (joined.unwrap_or_default(), "string".into(), None)
+        }
+
+        // [Name, Description, UpperBoundValue, LowerBoundValue, Unit, SetPointValue]
+        "IFCPROPERTYBOUNDEDVALUE" => {
+            let upper = entity.get(2).and_then(|v| v.as_float());
+            let lower = entity.get(3).and_then(|v| v.as_float());
+            let set_point = entity.get(5).and_then(|v| v.as_float());
+            let display_value = set_point.or(upper).or(lower);
+            match display_value {
+                None => (String::new(), "null".into(), None),
+                Some(dv) => {
+                    let mut display = fmt_number(dv);
+                    if let (Some(lo), Some(hi)) = (lower, upper) {
+                        // en-dash to match the WASM display exactly.
+                        display.push_str(&format!(
+                            " [{} \u{2013} {}]",
+                            fmt_number(lo),
+                            fmt_number(hi)
+                        ));
+                    }
+                    let data_type = infer_data_type(entity.get(5))
+                        .or_else(|| infer_data_type(entity.get(2)))
+                        .or_else(|| infer_data_type(entity.get(3)));
+                    // The value is a display STRING ("5 [2 – 8]"), so keep kind
+                    // `string` — a `real` kind would make the client `Number()`
+                    // it to NaN. The Lists cell derives from the value, and the
+                    // measure tag still rides on `data_type`.
+                    (display, "string".into(), data_type)
+                }
+            }
+        }
+
+        // [Name, Description, DefiningValues (list), DefinedValues (list), ...]
+        "IFCPROPERTYTABLEVALUE" => {
+            let rows = match entity.get(2) {
+                Some(AttributeValue::List(items)) => items.len(),
+                _ => 0,
+            };
+            if rows > 0 {
+                (format!("Table ({} rows)", rows), "string".into(), None)
+            } else {
+                (String::new(), "null".into(), None)
+            }
+        }
+
+        // [Name, Description, PropertyReference]
+        "IFCPROPERTYREFERENCEVALUE" => match entity.get(2).and_then(|v| v.as_entity_ref()) {
+            Some(id) => (format!("#{}", id), "string".into(), None),
+            None => (String::new(), "null".into(), None),
+        },
+
+        _ => return None,
+    };
+
     Some(Property {
         property_name,
         property_value,
@@ -86,10 +152,65 @@ fn extract_property(entity: &DecodedEntity, _decoder: &mut EntityDecoder) -> Opt
     })
 }
 
+/// Stringify one member of an enumerated / list value, mirroring the WASM
+/// `String(v)` / `String(v[1])`: a typed wrapper `List([type, inner])` yields
+/// the inner, a scalar yields itself. `None` for nulls (dropped, like the TS
+/// `.filter(v => v !== 'null')`).
+fn stringify_member(v: &ifc_lite_core::AttributeValue) -> Option<String> {
+    use ifc_lite_core::AttributeValue;
+    match v {
+        AttributeValue::List(items) if items.len() == 2 => stringify_scalar(&items[1]),
+        other => stringify_scalar(other),
+    }
+}
+
+fn stringify_scalar(v: &ifc_lite_core::AttributeValue) -> Option<String> {
+    use ifc_lite_core::AttributeValue;
+    match v {
+        AttributeValue::String(s) => Some(s.clone()),
+        AttributeValue::Enum(e) => Some(e.clone()),
+        AttributeValue::Integer(i) => Some(i.to_string()),
+        AttributeValue::Float(f) => Some(fmt_number(*f)),
+        _ => None,
+    }
+}
+
+/// Join a `List(...)` attribute's members with ", " (WASM `values.join(', ')`),
+/// or `None` when the attribute is not a non-empty list.
+fn join_member_list(attr: Option<&ifc_lite_core::AttributeValue>) -> Option<String> {
+    use ifc_lite_core::AttributeValue;
+    match attr {
+        Some(AttributeValue::List(items)) => {
+            let parts: Vec<String> = items.iter().filter_map(stringify_member).collect();
+            if parts.is_empty() {
+                None
+            } else {
+                Some(parts.join(", "))
+            }
+        }
+        _ => None,
+    }
+}
+
+/// The IFC type tag of a typed-wrapper attribute (`List([String(type), _])`),
+/// upper-cased — mirrors the WASM `inferDataType`.
+fn infer_data_type(attr: Option<&ifc_lite_core::AttributeValue>) -> Option<String> {
+    use ifc_lite_core::AttributeValue;
+    match attr {
+        Some(AttributeValue::List(items)) if items.len() == 2 => match &items[0] {
+            AttributeValue::String(s) => Some(s.to_uppercase()),
+            _ => None,
+        },
+        _ => None,
+    }
+}
+
 /// Normalise an `.ENUM.`-style logical/boolean token (dots already stripped to
 /// the inner letter by the tokenizer, e.g. `Enum("T")`; or a bare `String`).
 fn logical_token(v: &ifc_lite_core::AttributeValue) -> Option<&str> {
-    v.as_enum().or_else(|| v.as_string()).map(|s| s.trim_matches('.'))
+    v.as_enum()
+        .or_else(|| v.as_string())
+        .map(|s| s.trim_matches('.'))
 }
 
 /// Resolve an `IfcPropertySingleValue` NominalValue → (value string, kind, data_type),
